+++ src/components/Nav/Nav.scss
@@ -83,7 +83,7 @@
           width: 28px;
           height: 28px;
           text-indent: -9999px;
-          background: url(/assets/nav/basketIcon.png) no-repeat;
+          background: url(/assets/Nav/basketIcon.png) no-repeat;
           background-size: 28px;
 
           .num {
