+++ src/pages/ProductDetail/Leftside.scss
@@ -1,0 +1,185 @@
+.left-side {
+  margin-left: 50px;
+  width: 700px;
+  max-width: 800px;
+  display: inline-block;
+
+  .detail-box {
+    font-size: 0px;
+    display: flex;
+    margin-bottom: 80px;
+
+    .detail-img {
+      width: 310px;
+      min-width: 310px;
+      opacity: 1;
+      transform: translate3d(0px, 0px, 0px);
+    }
+    .detailImg {
+      outline: none;
+      width: 100%;
+    }
+    .detail-string {
+      margin-left: 20px;
+      padding: 0px;
+      width: 100%;
+      height: 100%;
+      flex-direction: column;
+      font-size: 20px;
+      font-weight: bold;
+      color: #707070;
+
+      .sector-first {
+        line-height: 1.11;
+        text-align: left;
+
+        .detail-mainString {
+          font-size: 24px;
+          overflow: hidden;
+          text-overflow: ellipsis;
+          max-width: 27ch;
+          margin-bottom: 10px;
+        }
+      }
+
+      .sector-second {
+        line-height: 1.11;
+        text-align: left;
+
+        .detail-secondString {
+          font-size: 16px;
+          color: rgb(178, 178, 178);
+          margin-bottom: 10px;
+          font-weight: bold;
+        }
+
+        .detail-hashtag {
+          font-size: 16px;
+          color: rgb(255, 186, 0);
+          margin-bottom: 10px;
+        }
+
+        .rating {
+          display: flex;
+          align-items: flex-start;
+
+          .rating-bg {
+            width: 100px;
+            height: 20px;
+            background: url(/assets/ProductDetail/icon-star-score-bg.png)
+              no-repeat;
+            background-size: 100px 20px;
+
+            .rating {
+              height: 20px;
+              background: url(/assets/ProductDetail/icon-star-score.png)
+                no-repeat;
+              background-size: 100px 20px;
+            }
+          }
+
+          .review {
+            margin-left: 10px;
+            margin-bottom: -8px;
+            text-decoration: none;
+            font-size: 14px;
+            color: rgb(0, 151, 243);
+            margin-bottom: 50px;
+          }
+        }
+      }
+
+      .sector-third {
+        line-height: 1.14;
+        margin-bottom: 50px;
+
+        .flex {
+          display: flex;
+          -webkit-box-align: center;
+          align-items: center;
+          color: rgb(112, 112, 112);
+          margin-bottom: 10px;
+          word-break: keep-all;
+        }
+
+        .third-string {
+          font-size: 14px;
+          font-weight: bold;
+        }
+
+        .delivery_string {
+          font-size: 15px;
+          font-weight: bold;
+          color: #b2b2b2;
+        }
+      }
+
+      .sector-fourth {
+        margin-bottom: 0px;
+        color: rgb(178, 178, 178);
+
+        .flex-price {
+          font-size: 14px;
+          font-weight: bold;
+          color: #707070;
+        }
+
+        .discountBox {
+          margin-top: 10px;
+          font-weight: bold;
+          font-size: 36px;
+          color: #707070;
+        }
+
+        .fourth-div {
+          // margin-top: 10px;
+        }
+
+        .fourth-string {
+          font-size: 14px;
+          font-weight: bold;
+          color: rgb(0, 151, 243);
+        }
+      }
+    }
+  }
+
+  .image-stack {
+    height: 600px;
+    overflow: hidden;
+    position: relative;
+
+    &.active {
+      height: auto;
+    }
+
+    .foled-button-down {
+      position: absolute;
+      bottom: 30px;
+      left: 50%;
+      transform: translateX(-50%);
+
+      .fold-btn {
+        width: 600px;
+        height: 44px;
+        border-radius: 5px;
+        display: flex;
+
+        justify-content: center;
+
+        align-items: center;
+        gap: 4.42px;
+        background-color: rgb(255, 186, 0);
+        text-align: center;
+        font-weight: 500;
+        font-size: 18px;
+        line-height: 100%;
+        color: rgb(255, 255, 255);
+      }
+    }
+  }
+
+  .recommend {
+    margin-top: 40px;
+  }
+}
